+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [server] Implement global filtering capabilities, to selectively advertize globals to clients.
   - **Breaking**: this bumps the minimal version of the C libraries to 1.13.0
 - [server] **Breaking**: migrate the event loop logic to using `calloop` rather than our own
+- [client] Add the `eventloop` cargo feature which provides `calloop` integration for `EventQueue`
 
 # 0.21.0-alpha1 - 2018-07-18